Fixing compile errors pt 4

+++ grpc-client/src/main/java/com/example/grpc/client/grpcclient/GRPCClientService.java
@@ -211,7 +211,7 @@ public class GRPCClientService {
 		return;
 	}
 
-	static double getDeadline(Matrix A1, Matrix A2, MatrixServiceBlockingStub stub, int numberOfBlocks, double deadline) {
+	static int getDeadline(Matrix A1, Matrix A2, MatrixServiceBlockingStub stub, int numberOfBlocks, double deadline) {
 		int deadlineMilis = (int) (deadline * 1000);
 		double startTime = System.currentTimeMillis();
 		MatrixResponse temp = stub.multiplyMatrix(MatrixRequest.newBuilder().setA(A1).setB(A2).build());
